Made the module add the vertex of the oldest feed item first to have all feed items inserted into the database in a continuous order in the databas from the oldest to the newest feed.
Also made the publish-date use the imported date if no date was specified in the feed.

Lastly made the feed vertexes only add new data when such exists (like on an updated feed item) and not replace the current feed data each and every time the feed is fetched (as today). The update now only works on feed items with a valid publish date since otherwise 0 is used to decide when a feed items version is valid and not and we can't identify a change.